Crate axdriver_display

Common traits and types for graphics display device drivers.

Re-exports

pub use axdriver_base::BaseDriverOps;

pub use axdriver_base::DevError;

pub use axdriver_base::DevResult;

pub use axdriver_base::DeviceType;

Structs

DisplayInfo

The information of the graphics device.

FrameBuffer

The framebuffer.

Traits

DisplayDriverOps

Operations that require a graphics device driver to implement.