axdriver_display

Trait DisplayDriverOps

pub trait DisplayDriverOps: BaseDriverOps {
    // Required methods
    fn info(&self) -> DisplayInfo;
    fn fb(&self) -> FrameBuffer<'_>;
    fn need_flush(&self) -> bool;
    fn flush(&mut self) -> DevResult;
}

Operations that require a graphics device driver to implement.

Required Methods

fn info(&self) -> DisplayInfo

Get the display information.

fn fb(&self) -> FrameBuffer<'_>

Get the framebuffer.

fn need_flush(&self) -> bool

Whether need to flush the framebuffer to the screen.

fn flush(&mut self) -> DevResult

Flush framebuffer to the screen.